+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Conway/Ledger.hs
@@ -127,7 +127,7 @@ instance
 
   environmentSpec ConwayLedgerExecContext {..} =
     let UtxoExecContext {..} = clecUtxoExecContext
-     in constrained' $ \slotNo _ _txIx pp _acntSt _mempool ->
+     in constrained' $ \slotNo _ _txIx pp _acntSt ->
           [ assert $ pp ==. lit (uePParams uecUtxoEnv)
           , assert $ slotNo ==. lit (ueSlot uecUtxoEnv)
           ]